requirements.txt
numpy
PyYAML
mujoco
matplotlib

# Optional (only if you want to run Julia action selection)
julia

